pyproject.toml
[tool.poetry]
name = "fastapi-architecture"
version = "0.1.0"
description = ""
authors = ["Hongmin <nhm9001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.0"
aiomysql = "^0.2.0"
redis = "^5.2.1"
dependency-injector = "^4.44.0"
pyyaml = "^6.0.2"
ujson = "^5.10.0"
jwt = "^1.3.1"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"
alembic = "^1.14.0"
coverage = "^7.6.10"
pyjwt = "^2.10.1"
httpx = "^0.28.1"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
pytest-dotenv = "^0.5.2"
pytest-html = "^4.1.1"
numpy = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
